requirements.txt
python-dotenv
pandas 
numpy
seaborn
matplotlib
scikit-learn
catboost
xgboost
dill
flask
pymongo
pymongo[srv]==3.6
certifi
pyaml



#-e .



